Replace Approve warning with sign-in button when metamask is not connected
