+++ src/components/Header/SignInButton.js
@@ -1,31 +1,12 @@
 import React from 'react'
-import styled from 'react-emotion'
-import { Link } from 'react-router-dom'
 
 import { GlobalConsumer } from '../../GlobalState'
 import Tooltip from '../Tooltip'
 import Button from '../Forms/Button'
-import Avatar from '../User/Avatar'
 import { CANNOT_RESOLVE_ACCOUNT_ADDRESS } from '../../utils/errors'
 import Assist from './Assist'
 
-const Account = styled(Link)`
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-`
-const Username = styled('div')`
-  max-width: 100px;
-  color: white;
-  font-family: 'Muli';
-  margin-right: 5px;
-  font-size: 16px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-`
-
-function SignInButton() {
+const SignInButton = ({ children, type, ...props }) => {
   const _signIn = ({
     showTooltip,
     hideTooltip,
@@ -50,55 +31,27 @@ function SignInButton() {
 
   return (
     <GlobalConsumer>
-      {({
-        reloadUserAddress,
-        userProfile,
-        networkState,
-        loggedIn,
-        signIn,
-        showModal
-      }) => {
-        const twitterProfile =
-          userProfile &&
-          userProfile.social &&
-          userProfile.social.find(s => s.type === 'twitter')
-        return loggedIn && userProfile ? (
-          <>
-            {/* <Notifications>Notification</Notifications> */}
-            <Account to={`/user/${userProfile.username}`}>
-              {userProfile ? (
-                <Username data-testid="userprofile-name">
-                  {userProfile.username}
-                </Username>
-              ) : null}
-              <Avatar
-                src={`https://avatars.io/twitter/${
-                  twitterProfile ? twitterProfile.value : 'unknowntwitter123abc'
-                }/medium`}
-              />
-            </Account>
-          </>
-        ) : (
-          <Tooltip text={CANNOT_RESOLVE_ACCOUNT_ADDRESS} position="left">
-            {({ tooltipElement, showTooltip, hideTooltip }) => (
-              <Button
-                type="light"
-                onClick={_signIn({
-                  showTooltip,
-                  hideTooltip,
-                  signIn,
-                  reloadUserAddress,
-                  networkState
-                })}
-                analyticsId="Sign In"
-              >
-                {tooltipElement}
-                Sign in
-              </Button>
-            )}
-          </Tooltip>
-        )
-      }}
+      {({ reloadUserAddress, networkState, signIn, showModal }) => (
+        <Tooltip text={CANNOT_RESOLVE_ACCOUNT_ADDRESS} position="left">
+          {({ tooltipElement, showTooltip, hideTooltip }) => (
+            <Button
+              {...props}
+              type={type}
+              onClick={_signIn({
+                showTooltip,
+                hideTooltip,
+                signIn,
+                reloadUserAddress,
+                networkState
+              })}
+              analyticsId="Sign In"
+            >
+              {tooltipElement}
+              {children || 'Sign In'}
+            </Button>
+          )}
+        </Tooltip>
+      )}
     </GlobalConsumer>
   )
 }
